Move ContractConfigDeserializer to separate file
