+++ core/src/main/kotlin/io/specmatic/core/config/v2/ContractConfig.kt
@@ -1,15 +1,11 @@
 package io.specmatic.core.config.v2
 
 import com.fasterxml.jackson.core.JsonGenerator
-import com.fasterxml.jackson.core.JsonParser
-import com.fasterxml.jackson.databind.DeserializationContext
-import com.fasterxml.jackson.databind.JsonDeserializer
-import com.fasterxml.jackson.databind.JsonMappingException
-import com.fasterxml.jackson.databind.JsonNode
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize
 import com.fasterxml.jackson.databind.annotation.JsonSerialize
 import io.specmatic.core.Source
 import io.specmatic.core.SourceProvider
+import io.specmatic.core.config.ContractConfigDeserializer
 import io.specmatic.core.config.ContractConfigSerializer
 
 @JsonSerialize(using = ContractConfigSerializer::class)
@@ -87,38 +83,5 @@ data class ContractConfig(
         override fun isEmpty(): Boolean {
             return directory == "."
         }
-    }
-}
-
-class ContractConfigDeserializer : JsonDeserializer<ContractConfig>() {
-    override fun deserialize(parser: JsonParser, ctxt: DeserializationContext): ContractConfig {
-        val node: JsonNode = parser.codec.readTree(parser)
-
-        val contractSource = when {
-            node.has("git") -> {
-                val gitNode = node.get("git")
-                ContractConfig.GitContractSource(
-                    url = gitNode.get("url").asText(),
-                    branch = gitNode.get("branch").asText()
-                )
-            }
-
-            node.has("filesystem") -> {
-                val filesystemNode = node.get("filesystem")
-                ContractConfig.FileSystemContractSource(
-                    directory = filesystemNode.get("directory").asText()
-                )
-            }
-
-            else -> throw JsonMappingException.from(
-                parser,
-                "Contracts field must have either 'git' or 'filesystem' field"
-            )
-        }
-
-        val provides = node.get("provides").map { it.asText() }
-        val consumes = node.get("consumes").map { it.asText() }
-
-        return ContractConfig(contractSource, provides, consumes)
     }
 }
